+++ android/engine/src/main/java/org/smartregister/fhircore/engine/util/extension/PatientExtension.kt
@@ -17,17 +17,12 @@
 package org.smartregister.fhircore.engine.util.extension
 
 import android.content.Context
-import java.text.SimpleDateFormat
 import java.time.LocalDate
 import java.time.Period
 import java.time.ZoneId
 import java.time.temporal.ChronoUnit
 import java.util.Date
-import java.util.Locale
-import org.hl7.fhir.r4.model.Condition
 import org.hl7.fhir.r4.model.Enumerations
-import org.hl7.fhir.r4.model.Identifier
-import org.hl7.fhir.r4.model.Immunization
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.codesystems.AdministrativeGender
 import org.smartregister.fhircore.engine.R
@@ -36,26 +31,6 @@ private const val RISK = "risk"
 const val DAYS_IN_YEAR = 365
 const val DAYS_IN_MONTH = 30
 const val DAYS_IN_WEEK = 7
-
-fun Patient.extractName(): String {
-  if (!hasName()) return ""
-  val humanName = this.name.firstOrNull()
-  return if (humanName != null) {
-    (humanName.given + humanName.family).filterNotNull().joinToString(" ") {
-      it.toString().trim().capitalizeFirstLetter()
-    }
-  } else ""
-}
-
-fun Patient.extractFamilyName(): String {
-  if (!hasName()) return ""
-  val humanName = this.name.firstOrNull()
-  return if (humanName != null) {
-    humanName.family?.capitalizeFirstLetter()?.plus(" Family") ?: ""
-  } else ""
-}
-
-fun String.capitalizeFirstLetter() = replaceFirstChar { it.titlecase(Locale.getDefault()) }
 
 fun Patient.extractGender(context: Context): String? =
   if (hasGender()) {
@@ -113,68 +88,6 @@ fun getAgeStringFromDays(date: Date): String {
   } else elapseDaysString
 }
 
-fun Patient.atRisk() =
-  this.extension.singleOrNull { it.value.toString().contains(RISK) }?.value?.toString() ?: ""
-
-fun Patient.getLastSeen(immunizations: List<Immunization>): String {
-  return immunizations
-    .maxByOrNull { it.protocolAppliedFirstRep.doseNumberPositiveIntType.value }
-    ?.occurrenceDateTimeType
-    ?.toDisplay()
-    ?: this.meta?.lastUpdated.lastSeenFormat()
-}
-
-fun Date?.lastSeenFormat(): String {
-  return if (this != null) {
-    SimpleDateFormat("MM-dd-yyyy", Locale.ENGLISH).run { format(this@lastSeenFormat) }
-  } else ""
-}
-
-fun Patient.extractAddress(): String {
-  if (!hasAddress()) return ""
-  return with(addressFirstRep) {
-    val addressLine =
-      if (this.hasLine()) this.line.joinToString(separator = ", ", postfix = ", ") else ""
-
-    addressLine
-      .join(this.district, " ")
-      .join(this.city, " ")
-      .join(this.state, " ")
-      .join(this.country, " ")
-      .trim()
-  }
-}
-
-fun Patient.extractAddressDistrict(): String {
-  return with(addressFirstRep) { this.district ?: "" }
-}
-
-fun Patient.extractAddressState(): String {
-  return with(addressFirstRep) { this.state ?: "" }
-}
-
-fun Patient.extractAddressText(): String {
-  return with(addressFirstRep) { this.text ?: "" }
-}
-
-fun Patient.extractTelecom(): List<String>? {
-  if (!hasTelecom()) return null
-  return telecom.map { it.value }
-}
-
-fun Patient.extractGeneralPractitionerReference(): String {
-  if (!hasGeneralPractitioner()) return ""
-  return with(generalPractitionerFirstRep) { this.reference }
-}
-
-fun Patient.extractManagingOrganizationReference(): String {
-  if (!hasManagingOrganization()) return ""
-  return with(managingOrganization) { this.reference }
-}
-
-fun Patient.extractDeathDate() =
-  if (this.hasDeceasedDateTimeType()) deceasedDateTimeType?.value else null
-
 fun String?.join(other: String?, separator: String) =
   this.orEmpty().plus(other?.plus(separator).orEmpty())
 
@@ -183,41 +96,9 @@ fun Patient.extractFamilyTag() =
     it.display.contentEquals("family", true) || it.display.contains("head", true)
   }
 
-fun Patient.isFamilyHead() = this.extractFamilyTag() != null
-
-fun List<Condition>.hasActivePregnancy() =
-  this.any { condition ->
-    // is active and any of the display / text into code is pregnant
-    val active = condition.clinicalStatus.coding.any { it.code == "active" }
-    val pregnancy =
-      condition.code.coding.map { it.display }.plus(condition.code.text).any {
-        it.contentEquals("pregnant", true)
-      }
-
-    active && pregnancy
-  }
-
-fun List<Condition>.pregnancyCondition(): Condition {
-  var pregnancyCondition = Condition()
-  this.forEach { condition ->
-    if (condition.code.coding.map { it.display }.plus(condition.code.text).any {
-        it.contentEquals("pregnant", true)
-      }
-    )
-      pregnancyCondition = condition
-  }
-
-  return pregnancyCondition
-}
-
 fun Enumerations.AdministrativeGender.translateGender(context: Context) =
   when (this) {
     Enumerations.AdministrativeGender.MALE -> context.getString(R.string.male)
     Enumerations.AdministrativeGender.FEMALE -> context.getString(R.string.female)
     else -> context.getString(R.string.unknown)
   }
-
-fun Patient.extractOfficialIdentifier(): String? =
-  if (this.hasIdentifier())
-    this.identifier.firstOrNull { it.use == Identifier.IdentifierUse.OFFICIAL }?.value
-  else null
